+++ src/Calvano2020/summary.jl
@@ -4,9 +4,9 @@ using ReinforcementLearning
 profit_measure(π_hat::Vector{Float64}, π_N, π_M) = (mean(π_hat) - π_N) / (π_M - π_N)
 
 struct CalvanoSummary
-    α::Float64
-    β::Float64
-    avg_profit::Vector{Float64}
+    α::Float32
+    β::Float32
+    avg_profit::Vector{Float32}
 end
 
 function economic_summary(e::ReinforcementLearningCore.Experiment)
@@ -15,7 +15,7 @@ function economic_summary(e::ReinforcementLearningCore.Experiment)
     π_N = e.env.env.profit_function(fill(e.env.env.p_Bert_nash_equilibrium, 2))[1]
     π_M = e.env.env.profit_function(fill(e.env.env.p_monop_opt, 2))[1]
 
-    avg_profit = Float64[]
+    avg_profit = Float32[]
 
     for i in [1, 2]
         @chain e.hook.hooks[i][1].rewards[(end-convergence_threshold):end] begin
